fix: invalid auth check

+++ backend/src/main/java/format/backend/submission/service/SubmissionService.java
@@ -133,7 +133,7 @@ public class SubmissionService {
                 .map(a -> a.getId().equals(keycloakJwtClaims.sub()))
                 .orElse(false);
         val isAdmin = keycloakJwtClaims.roles().contains(Role.ADMIN);
-        if (!isFormOwner && !isAdmin) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        if (!(isFormOwner || isAdmin)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
 
         submissionRepository.delete(submission);
         formService.decrementSubmissionsCountById(form.getId());
